调整交易分析接口

+++ src/main/java/com/fx/xzt/sys/service/OrderAnalysisService.java
@@ -15,14 +15,14 @@ import java.util.Map;
 public interface OrderAnalysisService {
 
     PageInfo<Map<String, Object>> orderAnalysis(String startTime, String endTime, String agentName,
-                                                Short upOrDown,Short orderState, Short profitLoss,Long agentId,
+                                                Integer upOrDown,Integer orderState, Integer profitLoss,Long agentId,
                                                 Integer pageNum, Integer pageSize) throws ParseException;
     Map orderAnalysisCount(String startTime, String endTime,
-                                                 Short upOrDown,Short orderState, Short profitLoss,Long agentId,
+                           Integer upOrDown,Integer orderState, Integer profitLoss,Long agentId,
                                                  Integer pageNum, Integer pageSize);
 
     List<Map<String, Object>> exportAnalysis(String startTime, String endTime, String agentName,
-                                                Short upOrDown,Short orderState, Short profitLoss,Long agentId
+                                             Integer upOrDown,Integer orderState, Integer profitLoss,Long agentId
                                                 );
